+++ src/mmw/js/src/analyze/models.js
@@ -26,25 +26,9 @@ var AnalyzeTaskModel = coreModels.TaskModel.extend({
     )
 });
 
-var AnalyzeMessageModel = Backbone.Model.extend({
-    message: null,
-    iconClass: null,
-
-    setError: function() {
-        this.set('message', 'Error');
-        this.set('iconClasses', 'fa fa-exclamation-triangle');
-    },
-
-    setAnalyzing: function() {
-        this.set('message', 'Analyzing');
-        this.set('iconClasses', 'fa fa-circle-o-notch fa-spin');
-    }
-});
-
 module.exports = {
     AnalyzeTaskModel: AnalyzeTaskModel,
     LayerModel: LayerModel,
     LayerCollection: LayerCollection,
     LayerCategoryCollection: LayerCategoryCollection,
-    AnalyzeMessageModel: AnalyzeMessageModel
 };
