+++ src/mmw/js/src/core/templates/message.html
